Transform license steps into check license task.
