+++ lib/cli/commands/license.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const globalOptionDefinitions = require('../globalOptionDefinitions'),
-      { checkLicenseCompatibility: checkLicenseCompatibilityStep } = require('../../steps/license'),
+const checkLicenseCompatibilityStep = require('../tasks/checkLicenseCompatibility'),
+      globalOptionDefinitions = require('../globalOptionDefinitions'),
       ui = require('../ui');
 
 const license = {
@@ -27,22 +27,7 @@ const license = {
       ]);
     }
 
-    try {
-      await checkLicenseCompatibilityStep({ directory, ui });
-
-      ui.printTaskSuccess('License check successful.');
-    } catch (ex) {
-      switch (ex.code) {
-        case 'ELICENSEINCOMPATIBLE':
-        case 'ELICENSENOTSUPPORTED':
-          ui.printTaskFailure('License check failed.');
-          break;
-        default:
-          ui.error('Failed to run license check.');
-      }
-
-      throw ex;
-    }
+    await checkLicenseCompatibilityStep({ directory, ui });
   }
 };
