+++ lib/cli/commands/qa.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const analyseTask = require('../tasks/analyse'),
+      checkLicenseCompatibilityTask = require('../tasks/checkLicenseCompatibility'),
       dependenciesTask = require('../tasks/dependencies'),
       globalOptionDefinitions = require('../globalOptionDefinitions'),
-      licenseSteps = require('../../steps/license'),
       testSteps = require('../../steps/test'),
       ui = require('../ui');
 
@@ -51,19 +51,7 @@ const deps = {
     }
 
     await dependenciesTask({ directory, ui });
-
-    try {
-      await licenseSteps.checkLicenseCompatibility({ directory, ui });
-    } catch (ex) {
-      switch (ex.code) {
-        case 'ELICENSEINCOMPATIBLE':
-          throw ex;
-        case 'ELICENSENOTSUPPORTED':
-          break;
-        default:
-          throw ex;
-      }
-    }
+    await checkLicenseCompatibilityTask({ directory, ui });
 
     ui.printTaskSuccess('Quality assurance successful.');
   }
