+++ lib/cli/commands/release.js
@@ -5,10 +5,10 @@ const path = require('path');
 const analyseTask = require('../tasks/analyse'),
       buildTask = require('../tasks/build'),
       { bumpVersion: bumpVersionStep, generateTocForMarkdown: generateTocForMarkdownStep, git: gitSteps } = require('../../steps/release'),
+      checkLicenseCompatibilityTask = require('../tasks/checkLicenseCompatibility'),
       dependenciesTask = require('../tasks/dependencies'),
       files = require('../../files'),
       globalOptionDefinitions = require('../globalOptionDefinitions'),
-      licenseSteps = require('../../steps/license'),
       testSteps = require('../../steps/test'),
       ui = require('../ui');
 
@@ -78,19 +78,7 @@ const release = {
         }
 
         await dependenciesTask({ directory, ui });
-
-        try {
-          await licenseSteps.checkLicenseCompatibility({ directory, ui });
-        } catch (ex) {
-          switch (ex.code) {
-            case 'ELICENSEINCOMPATIBLE':
-              throw ex;
-            case 'ELICENSENOTSUPPORTED':
-              break;
-            default:
-              throw ex;
-          }
-        }
+        await checkLicenseCompatibilityTask({ directory, ui });
       }
 
       await gitSteps.checkIfOnMaster({ directory, ui });
