+++ lib/errors.js
@@ -10,6 +10,7 @@ const errors = defekt({
   DirectoryAccessFailed: {},
   ExecutableFailed: {},
   FileParsingFailed: {},
+  LicenseCheckFailed: {},
   LicenseIncompatible: {},
   LicenseNotSupported: {},
   LocalMasterBehindRemote: {},
