+++ lib/steps/license/index.js
@@ -1,13 +1,11 @@
 'use strict';
 
-const checkLicenseCompatibility = require('./checkLicenseCompatibility'),
-      findIncompatiblePackages = require('./findIncompatiblePackages'),
+const findIncompatiblePackages = require('./findIncompatiblePackages'),
       getLicense = require('./getLicense'),
       getLicenseFromWhiteList = require('./getLicenseFromWhiteList'),
       getWhitelistedPackages = require('./getWhitelistedPackages');
 
 module.exports = {
-  checkLicenseCompatibility,
   findIncompatiblePackages,
   getLicense,
   getLicenseFromWhiteList,
